image: Account for differences with mmdebstrap
Only legacy debootstrap needed moving of sources-list. With mmdebstrap,
we are now creating the list twice, and apt will complain about that on
the target.

Furthermore, the cleaning prior to installing custom lists needs to take
the different filenames with mmdebstrap into account.

Fixes: d0fd18ef87d2 ("meta: Allow selecting bootstrap providers")

+++ meta/classes/image.bbclass
@@ -416,12 +416,16 @@ do_rootfs_finalize() {
                 -maxdepth 1 -name 'qemu-*-static' -type f -delete
         fi
 
-        if [ -e "${ROOTFSDIR}/etc/apt/sources-list" ] && [ -d "${ROOTFSDIR}/etc/apt/sources.list.d" ]; then
+        # needed only for debootstrap, mmdebstrap leave a 0000bootstrap.list behind
+        if [ -e "${ROOTFSDIR}/etc/apt/sources-list" ] && \
+           [ -d "${ROOTFSDIR}/etc/apt/sources.list.d" ] && \
+           [ -z "$(find ${ROOTFSDIR}/etc/apt/sources.list.d -mindepth 1)" ]; then
             mv "${ROOTFSDIR}/etc/apt/sources-list" \
                 "${ROOTFSDIR}/etc/apt/sources.list.d/bootstrap.list"
         fi
+
         if [ -n "${IMAGE_LISTS}" ]; then
-            rm -f "${ROOTFSDIR}/etc/apt/sources.list.d/bootstrap.list"
+            find "${ROOTFSDIR}/etc/apt/sources.list.d/" ! -type d -exec rm -f {} \;
             for l in ${IMAGE_LISTS}; do
                 cp "${WORKDIR}"/${l} "${ROOTFSDIR}/etc/apt/sources.list.d/"
             done
